intermediate requirements ready!

+++ src/main/java/interfaces/MixtureInfo.java
@@ -3,7 +3,7 @@ package interfaces;
 
 public interface MixtureInfo {
     
-    float calculateCalories();
+   double calculateCalories();
     String getColor();
     
 }
